Make Chosen.JS look better.
It is easy to use this feature in a single module.
But I think it will be very hard to implement this effect in the core.
We really need some rules set for individual themes, because a theme should not
overwrite margin and padding declarations for UL/LI with a higher weight.
(e.g. SeaBreeze theme)